Split native ABI test and record helpers
